requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
python-multipart
jinja2
pathlib
numpy
easyocr
paddleocr
paddlepaddle
webcolors
scikit-learn
boto3
python-dotenv
sqlalchemy
psycopg2-binary
